Revert "BODS-9368: Changes to historic matching script to improve cpu usage" (#688)
* Revert "BODS-9368: Changes to historic matching script to improve cpu usage"

* including change to how prod build is done

* Add original changeSets for historic data exports

* Fix formatting in db.changelog.xml for SQL file path

---------
