+++ liquibase/db.changelog.xml
@@ -495,12 +495,12 @@
         <sqlFile path="procedures/populate_avl_recorded_expected_journeys.sql" splitStatements="false"/>
     </changeSet>
     <changeSet id="158" author="abodsuser">
-        <sqlFile path="sql/158-replace-noc-adminarea-view.sql" splitStatements="false"/>
+        <sqlFile path="sql/158-replace-noc-adminarea-view.sql" splitStatements="false"/> 
     </changeSet>
     <changeSet id="159" author="abodsuser" runOnChange="true" runInTransaction="true">
         <sqlFile path="procedures/historic_subset_avl_export.sql" splitStatements="false"/>
     </changeSet>
     <changeSet id="160" author="abodsuser" runOnChange="true" runInTransaction="true">
         <sqlFile path="procedures/historic_subset_timetable_export.sql" splitStatements="false"/>
-    </changeSet>    
+    </changeSet>
 </databaseChangeLog>
